+++ embedded-graphics/src/drawable.rs
@@ -1,47 +1,6 @@
 //! `Drawable` trait and helpers
 use crate::{geometry::Point, pixelcolor::PixelColor, DrawTarget};
 
-/// A single pixel.
-///
-/// `Pixel` objects are used to specify the position and color of drawn pixels.
-///
-/// # Examples
-///
-/// The [`Drawable`] trait is implemented for `Pixel` which allows single pixels
-/// to be drawn to a [`DrawTarget`]:
-/// ```
-/// use embedded_graphics::{pixelcolor::BinaryColor, prelude::*};
-/// # use embedded_graphics::mock_display::MockDisplay;
-/// # let mut display = MockDisplay::new();
-///
-/// Pixel(Point::new(1, 2), BinaryColor::On).draw(&mut display);
-/// ```
-///
-/// Iterators with `Pixel` items can also be drawn:
-/// ```
-/// use embedded_graphics::{pixelcolor::BinaryColor, prelude::*};
-/// # use embedded_graphics::mock_display::MockDisplay;
-/// # let mut display = MockDisplay::new();
-///
-/// (0..100)
-///     .map(|i| Pixel(Point::new(i, i * 2), BinaryColor::On))
-///     .draw(&mut display);
-/// ```
-///
-/// [`Drawable`]: trait.Drawable.html
-/// [`DrawTarget`]: ../trait.DrawTarget.html
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct Pixel<C: PixelColor>(pub Point, pub C);
-
-impl<C> Drawable<C> for Pixel<C>
-where
-    C: PixelColor,
-{
-    fn draw<T: DrawTarget<C>>(self, display: &mut T) {
-        display.draw_pixel(self);
-    }
-}
-
 /// Marks an object as "drawable". Must be implemented for all graphics objects
 ///
 /// The `Drawable` trait describes how a particular graphical object is drawn. A `Drawable` object
@@ -70,7 +29,7 @@ where
 /// where
 ///     C: PixelColor + From<BinaryColor>,
 /// {
-///     fn draw<D: DrawTarget<C>>(self, display: &mut D) {
+///     fn draw<D: DrawTarget<C>>(self, display: &mut D) -> Result<(), D::Error> {
 ///         egrectangle!(
 ///             top_left = self.top_left,
 ///             bottom_right = self.bottom_right,
@@ -82,22 +41,22 @@ where
 ///             top_left = (20, 20),
 ///             style = text_style!(font = Font6x8, text_color = self.fg_color)
 ///         )
-///         .draw(display);
+///         .draw(display)
 ///     }
 /// }
 ///
-/// fn main() {
-///     let mut button = Button {
-///         top_left: Point::zero(),
-///         bottom_right: Point::new(100, 50),
-///         bg_color: Rgb888::RED,
-///         fg_color: Rgb888::BLUE,
-///         text: "Click me!",
-///     };
-///     # use embedded_graphics::mock_display::MockDisplay;
-///     # let mut display = MockDisplay::default();
-///     button.draw(&mut display);
-/// }
+/// let mut button = Button {
+///     top_left: Point::zero(),
+///     bottom_right: Point::new(100, 50),
+///     bg_color: Rgb888::RED,
+///     fg_color: Rgb888::BLUE,
+///     text: "Click me!",
+/// };
+///
+/// # use embedded_graphics::mock_display::MockDisplay;
+/// # let mut display = MockDisplay::default();
+/// button.draw(&mut display)?;
+/// # Ok::<(), core::convert::Infallible>(())
 /// ```
 ///
 /// [`DrawTarget`]: ../trait.DrawTarget.html
@@ -107,7 +66,51 @@ where
     C: PixelColor,
 {
     /// Draw the graphics object using the supplied DrawTarget.
-    fn draw<T: DrawTarget<C>>(self, display: &mut T);
+    fn draw<T: DrawTarget<C>>(self, display: &mut T) -> Result<(), T::Error>;
+}
+
+/// A single pixel.
+///
+/// `Pixel` objects are used to specify the position and color of drawn pixels.
+///
+/// # Examples
+///
+/// The [`Drawable`] trait is implemented for `Pixel` which allows single pixels
+/// to be drawn to a [`DrawTarget`]:
+/// ```
+/// use embedded_graphics::{pixelcolor::BinaryColor, prelude::*};
+/// # use embedded_graphics::mock_display::MockDisplay;
+/// # let mut display = MockDisplay::new();
+///
+/// Pixel(Point::new(1, 2), BinaryColor::On).draw(&mut display)?;
+/// # Ok::<(), core::convert::Infallible>(())
+/// ```
+///
+/// Iterators with `Pixel` items can also be drawn:
+///
+/// ```
+/// use embedded_graphics::{pixelcolor::BinaryColor, prelude::*};
+/// # use embedded_graphics::mock_display::MockDisplay;
+/// # let mut display = MockDisplay::new();
+///
+/// (0..100)
+///     .map(|i| Pixel(Point::new(i, i * 2), BinaryColor::On))
+///     .draw(&mut display)?;
+/// # Ok::<(), core::convert::Infallible>(())
+/// ```
+///
+/// [`Drawable`]: trait.Drawable.html
+/// [`DrawTarget`]: ../trait.DrawTarget.html
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub struct Pixel<C: PixelColor>(pub Point, pub C);
+
+impl<C> Drawable<C> for Pixel<C>
+where
+    C: PixelColor,
+{
+    fn draw<D: DrawTarget<C>>(self, display: &mut D) -> Result<(), D::Error> {
+        display.draw_pixel(self)
+    }
 }
 
 impl<C, T> Drawable<C> for &mut T
@@ -115,8 +118,8 @@ where
     C: PixelColor,
     T: Iterator<Item = Pixel<C>>,
 {
-    fn draw<D: DrawTarget<C>>(self, display: &mut D) {
-        display.draw_iter(self);
+    fn draw<D: DrawTarget<C>>(self, display: &mut D) -> Result<(), D::Error> {
+        display.draw_iter(self)
     }
 }
 
@@ -126,11 +129,11 @@ mod tests {
     use crate::{mock_display::MockDisplay, pixelcolor::BinaryColor};
 
     #[test]
-    fn draw_pixel() {
+    fn draw_pixel() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
-        Pixel(Point::new(0, 0), BinaryColor::On).draw(&mut display);
-        Pixel(Point::new(2, 1), BinaryColor::On).draw(&mut display);
-        Pixel(Point::new(1, 2), BinaryColor::On).draw(&mut display);
+        Pixel(Point::new(0, 0), BinaryColor::On).draw(&mut display)?;
+        Pixel(Point::new(2, 1), BinaryColor::On).draw(&mut display)?;
+        Pixel(Point::new(1, 2), BinaryColor::On).draw(&mut display)?;
 
         #[rustfmt::skip]
         assert_eq!(
@@ -141,5 +144,7 @@ mod tests {
             ]),
             display
         );
+
+        Ok(())
     }
 }