+++ embedded-graphics/src/fonts/font6x12.rs
@@ -81,11 +81,11 @@ mod tests {
     }
 
     #[test]
-    fn correct_m() {
+    fn correct_m() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new("Mm", Point::zero())
             .into_styled(TextStyle::new(Font6x12, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -104,14 +104,16 @@ mod tests {
                 "            ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn correct_ascii_borders() {
+    fn correct_ascii_borders() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new(" ~", Point::zero())
             .into_styled(TextStyle::new(Font6x12, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -130,14 +132,16 @@ mod tests {
                 "            ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn correct_dollar_y() {
+    fn correct_dollar_y() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new("$y", Point::zero())
             .into_styled(TextStyle::new(Font6x12, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -156,10 +160,12 @@ mod tests {
                 "        ##  ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn dont_panic() {
+    fn dont_panic() -> Result<(), core::convert::Infallible> {
         let two_question_marks = MockDisplay::from_pattern(&[
             "            ",
             "  ##    ##  ",
@@ -180,26 +186,28 @@ mod tests {
         let mut display = MockDisplay::new();
         Text::new("\0\n", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
 
         let mut display = MockDisplay::new();
         Text::new("\x7F\u{A0}", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
 
         let mut display = MockDisplay::new();
         Text::new("¡ÿ", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
 
         let mut display = MockDisplay::new();
         Text::new("Ā💣", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
+
+        Ok(())
     }
 
     #[test]