+++ embedded-graphics/src/fonts/font6x8.rs
@@ -84,11 +84,11 @@ mod tests {
     }
 
     #[test]
-    fn correct_m() {
+    fn correct_m() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new("Mm", Point::zero())
             .into_styled(TextStyle::new(Font6x8, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -103,10 +103,12 @@ mod tests {
                 "            ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn correct_inverse_colored_m() {
+    fn correct_inverse_colored_m() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         let style = TextStyle {
             font: Font6x8,
@@ -115,7 +117,7 @@ mod tests {
         };
         Text::new("Mm", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -130,11 +132,13 @@ mod tests {
                 "############",
             ])
         );
+
+        Ok(())
     }
 
     // tests if black on white has really the same behavior as white on black
     #[test]
-    fn compare_inverse_colored_m() {
+    fn compare_inverse_colored_m() -> Result<(), core::convert::Infallible> {
         let mut display_inverse = MockDisplay::new();
         let style_inverse = TextStyle {
             font: Font6x8,
@@ -143,7 +147,7 @@ mod tests {
         };
         Text::new("Mm", Point::zero())
             .into_styled(style_inverse)
-            .draw(&mut display_inverse);
+            .draw(&mut display_inverse)?;
 
         let mut display_normal = MockDisplay::new();
         let style_normal = TextStyle {
@@ -153,7 +157,7 @@ mod tests {
         };
         Text::new("Mm", Point::zero())
             .into_styled(style_normal)
-            .draw(&mut display_normal);
+            .draw(&mut display_normal)?;
 
         for y in 0..display_inverse.height() {
             for x in 0..display_inverse.width() {
@@ -165,14 +169,16 @@ mod tests {
                 assert_eq!(inverse_color, normal_color.map(|c| c.invert()));
             }
         }
+
+        Ok(())
     }
 
     #[test]
-    fn correct_ascii_borders() {
+    fn correct_ascii_borders() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new(" ~", Point::zero())
             .into_styled(TextStyle::new(Font6x8, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -187,24 +193,28 @@ mod tests {
                 "            ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn no_fill_doesnt_hang() {
+    fn no_fill_doesnt_hang() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new(" ", Point::zero())
             .into_styled(TextStyle::new(Font6x8, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(display, MockDisplay::new());
+
+        Ok(())
     }
 
     #[test]
-    fn correct_dollar_y() {
+    fn correct_dollar_y() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new("$y", Point::zero())
             .into_styled(TextStyle::new(Font6x8, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -219,14 +229,16 @@ mod tests {
                 "       ###  ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn correct_latin1() {
+    fn correct_latin1() -> Result<(), core::convert::Infallible> {
         let mut display = MockDisplay::new();
         Text::new("¡ÿ", Point::zero())
             .into_styled(TextStyle::new(Font6x8, BinaryColor::On))
-            .draw(&mut display);
+            .draw(&mut display)?;
 
         assert_eq!(
             display,
@@ -242,10 +254,12 @@ mod tests {
                 "            ",
             ])
         );
+
+        Ok(())
     }
 
     #[test]
-    fn dont_panic() {
+    fn dont_panic() -> Result<(), core::convert::Infallible> {
         let two_question_marks = MockDisplay::from_pattern(&[
             " ###   ### ",
             "#   # #   #",
@@ -261,19 +275,21 @@ mod tests {
         let mut display = MockDisplay::new();
         Text::new("\0\n", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
 
         let mut display = MockDisplay::new();
         Text::new("\x7F\u{A0}", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
 
         let mut display = MockDisplay::new();
         Text::new("Ā💣", Point::zero())
             .into_styled(style)
-            .draw(&mut display);
+            .draw(&mut display)?;
         assert_eq!(display, two_question_marks);
+
+        Ok(())
     }
 }