+++ embedded-graphics/tests/chaining.rs
@@ -25,7 +25,10 @@ impl From<u8> for TestPixelColor {
 }
 
 impl DrawTarget<TestPixelColor> for FakeDisplay {
-    fn draw_pixel(&mut self, _pixel: Pixel<TestPixelColor>) { // Noop
+    type Error = core::convert::Infallible;
+
+    fn draw_pixel(&mut self, _pixel: Pixel<TestPixelColor>) -> Result<(), Self::Error> {
+        Ok(())
     }
 
     fn size(&self) -> Size {
@@ -34,8 +37,8 @@ impl DrawTarget<TestPixelColor> for FakeDisplay {
 }
 
 #[test]
-fn it_supports_chaining() {
-    let mut disp = FakeDisplay {};
+fn it_supports_chaining() -> Result<(), core::convert::Infallible> {
+    let mut display = FakeDisplay {};
 
     let mut chained = Rectangle::new(Point::new(0, 0), Point::new(1, 1))
         .into_styled(PrimitiveStyle::default())
@@ -46,7 +49,7 @@ fn it_supports_chaining() {
                 .into_iter(),
         );
 
-    chained.draw(&mut disp);
+    chained.draw(&mut display)
 }
 
 fn multi() -> impl Iterator<Item = Pixel<TestPixelColor>> {
@@ -62,17 +65,17 @@ fn multi() -> impl Iterator<Item = Pixel<TestPixelColor>> {
 }
 
 #[test]
-fn return_from_fn() {
-    let mut disp = FakeDisplay {};
+fn return_from_fn() -> Result<(), core::convert::Infallible> {
+    let mut display = FakeDisplay {};
 
     let mut chained = multi();
 
-    chained.draw(&mut disp);
+    chained.draw(&mut display)
 }
 
 #[test]
-fn implicit_into_iter() {
-    let mut disp = FakeDisplay {};
+fn implicit_into_iter() -> Result<(), core::convert::Infallible> {
+    let mut display = FakeDisplay {};
 
     let mut chained = Rectangle::new(Point::new(0, 0), Point::new(1, 1))
         .into_styled(PrimitiveStyle::default())
@@ -83,5 +86,5 @@ fn implicit_into_iter() {
                 .into_iter(),
         );
 
-    chained.draw(&mut disp);
+    chained.draw(&mut display)
 }