+++ simulator/examples/fill-macros.rs
@@ -7,7 +7,7 @@ use embedded_graphics_simulator::{SimulatorDisplay, WindowBuilder};
 
 static CIRCLE_SIZE: i32 = 32;
 
-fn main() {
+fn main() -> Result<(), core::convert::Infallible> {
     let mut display = SimulatorDisplay::new(Size::new(384, 128));
 
     egcircle!(
@@ -15,7 +15,7 @@ fn main() {
         radius = CIRCLE_SIZE as u32,
         style = primitive_style!(stroke_color = BinaryColor::On, stroke_width = 1,)
     )
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egcircle!(
         center = (CIRCLE_SIZE, CIRCLE_SIZE),
@@ -27,7 +27,7 @@ fn main() {
         )
     )
     .translate(Point::new(16, 16))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egcircle!(
         center = (CIRCLE_SIZE, CIRCLE_SIZE),
@@ -39,7 +39,7 @@ fn main() {
         )
     )
     .translate(Point::new(CIRCLE_SIZE, CIRCLE_SIZE))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egrectangle!(
         top_left = (0, 0),
@@ -47,7 +47,7 @@ fn main() {
         style = primitive_style!(stroke_color = BinaryColor::On, stroke_width = 1,)
     )
     .translate(Point::new(96, 0))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egrectangle!(
         top_left = (0, 0),
@@ -59,7 +59,7 @@ fn main() {
         )
     )
     .translate(Point::new(96 + 16, 16))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egrectangle!(
         top_left = (0, 0),
@@ -71,14 +71,14 @@ fn main() {
         )
     )
     .translate(Point::new(96 + 32, 32))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egtriangle!(
         points = [(32, 0), (0, 64), (64, 64)],
         style = primitive_style!(stroke_color = BinaryColor::On, stroke_width = 1,)
     )
     .translate(Point::new(96 * 2, 0))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egtriangle!(
         points = [(32, 0), (0, 64), (64, 64)],
@@ -89,7 +89,7 @@ fn main() {
         )
     )
     .translate(Point::new(96 * 2 + 16, 16))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     egtriangle!(
         points = [(32, 0), (0, 64), (64, 64)],
@@ -100,11 +100,13 @@ fn main() {
         )
     )
     .translate(Point::new(96 * 2 + 32, 32))
-    .draw(&mut display);
+    .draw(&mut display)?;
 
     let mut window = WindowBuilder::new(&display)
         .title("Filled primitives using macros")
         .scale(2)
         .build();
     window.show_static(&display);
+
+    Ok(())
 }